Refactor: Split OpenSSH key handling into dedicated classes
- Rename PEMEncoder to OpenSSHKeyEncoder (it only handles OpenSSH format)
- Extract OpenSSH decoding from PEMDecoder into OpenSSHKeyDecoder
- Update PEMDecoder to delegate to OpenSSHKeyDecoder for OpenSSH keys
- Keep PEMDecoder for traditional PEM formats (RSA/DSA/EC private keys)

This separation makes the code organization clearer:
- OpenSSHKeyEncoder: exports keys in OpenSSH format (openssh-key-v1)
- OpenSSHKeyDecoder: imports keys in OpenSSH format
- PEMDecoder: handles traditional PEM formats, delegates OpenSSH to decoder

Note: Investigation revealed a bug in the original PEMDecoder padding check
(lines 666-672). The code used `tr.remain()` instead of `trEnc.remain()`,
but since `tr` is always empty at that point, padding was never verified.
For backward compatibility, we preserve this behavior. The checkInt matching
already validates successful decryption.
